Consolidate all profile logic
widget_tree can be made stateless
