+++ lib/mvp2/main/main_pages/widget_tree.dart
@@ -1,51 +1,24 @@
 import 'package:flutter/material.dart';
 
 import 'package:homebased_project/data/notifiers.dart';
-import 'package:homebased_project/backend/auth_api/auth_service.dart';
 import 'package:homebased_project/mvp2/main/main_components/main_seller_navbar_widget.dart';
 import 'package:homebased_project/mvp2/main/main_components/main_user_navbar_widget.dart';
-import 'package:homebased_project/mvp2/auth/auth_pages/auth_page.dart';
 import 'package:homebased_project/mvp2/profile/profile_pages/profile_app_bar_section.dart';
 
-class WidgetTree extends StatefulWidget {
+class WidgetTree extends StatelessWidget {
   const WidgetTree({super.key});
-
-  @override
-  State<WidgetTree> createState() => _WidgetTreeState();
-}
-
-class _WidgetTreeState extends State<WidgetTree> {
-  void _switchProfileMode() {
-    setState(() {
-      userMode.value = userMode.value == 'Seller' ? 'User' : 'Seller';
-    });
-    setUserMode(userMode.value);
-  }
-
-  void _logout() async {
-    await authService.signOut();
-    if (!mounted) return;
-    Navigator.pushReplacement(
-      context,
-      MaterialPageRoute(builder: (context) => const AuthPage()),
-    );
-  }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: ProfileAppBarSection(
-        profileMode: userMode.value,
-        onSwitchMode: _switchProfileMode,
-        onLogout: _logout,
-      ),
+      appBar: const ProfileAppBarSection(),
 
       body: ValueListenableBuilder<String>(
         valueListenable: userMode,
-        builder: (context, mode, child) {
+        builder: (context, mode, _) {
           return ValueListenableBuilder<int>(
             valueListenable: selectedPageNotifier,
-            builder: (context, selectedIndex, _) {
+            builder: (context, selectedIndex, __) {
               final pages = mode == "Seller" ? sellerPages : defaultUserPages;
               return pages.elementAt(selectedIndex);
             },
@@ -55,7 +28,7 @@ class _WidgetTreeState extends State<WidgetTree> {
 
       bottomNavigationBar: ValueListenableBuilder<String>(
         valueListenable: userMode,
-        builder: (context, mode, child) {
+        builder: (context, mode, _) {
           return mode == "Seller"
               ? const SellerNavbarWidget()
               : const DefaultUserNavbarWidget();
